requirements.txt
python-dotenv
fastapi
uvicorn[standard]
numpy
pandas
networkx
torch
causal-learn
google-generativeai
streamlit
matplotlib
requests